refactor: change stock and quantity types from float to integer across the codebase
- Updated the stock field in InventoryItemORM and AuditLogORM to use Integer instead of Float for better type accuracy.
- Modified the quantity field in StockUpdateInput and related functions to enforce integer values, ensuring stricter validation.
- Adjusted all relevant tests and functions to accommodate the new integer type for stock and quantity, enhancing consistency and reliability in inventory management.

+++ app/db/migrate.py
@@ -30,7 +30,7 @@ def load_inventory(db_url: str = "sqlite:///dental.db", json_path: Path = INVENT
                 id=raw["id"],
                 name=raw["name"],
                 category=raw["category"],
-                stock=float(raw["stock"]),
+                stock=int(raw["stock"]),
                 unit=raw["unit"],
                 flammable=bool(attrs.get("flammable", False)),
                 vasoconstrictor=bool(attrs.get("vasoconstrictor", False)),
